+++ src/IDListLayer.hpp
@@ -1,4 +1,5 @@
 #include <Geode/Geode.hpp>
+#include <Geode/utils/web.hpp>
 
 using namespace geode::prelude;
 
@@ -8,41 +9,36 @@ public:
     inline static std::vector<std::string> AREDL_NAMES = {};
     inline static std::vector<int> AREDL_POSITIONS = {};
     inline static bool AREDL_TRIED_LOADING = false;
-    static IDListLayer* create();
-    static void loadAREDL(bool fromMenuLayer) {
-        loadAREDL(fromMenuLayer, []() {});
-    }
-    static void loadAREDL(bool fromMenuLayer, MiniFunction<void()> callback);
+    inline static std::unordered_map<std::string, web::WebTask> RUNNING_REQUESTS = {};
 
-    void onClose(CCObject*);
-    void onSearch(CCObject*);
-    void onLeft(CCObject*);
-    void onRight(CCObject*);
-    void onInfo(CCObject*);
-    void onRefresh(CCObject*);
-    void onPage(CCObject*);
-    void onRandom(CCObject*);
-    void onFirst(CCObject*);
-    void onLast(CCObject*);
+    static IDListLayer* create();
+    static CCScene* scene();
+    template<class T>
+    static std::vector<T> pluck(matjson::Array const&, std::string const&);
+    static void loadAREDL(bool, MiniFunction<void()> callback = []() {});
+    static float createGap(CCNode*, CCNode*, float);
+
+    void search();
+    void page(int);
     void keyDown(enumKeyCodes) override;
     void keyBackClicked() override;
 
     ~IDListLayer();
 protected:
     GJListLayer* m_list;
-    CCMenu* m_backMenu;
-    CCMenu* m_leftMenu;
-    CCMenu* m_rightMenu;
-    CCMenu* m_infoMenu;
-    CCMenu* m_refreshMenu;
-    CCMenu* m_leftSearchMenu;
-    CCMenu* m_rightSearchMenu;
     CCLabelBMFont* m_listLabel;
     LoadingCircle* m_loadingCircle;
     CCLayerColor* m_searchBarView;
     CCTextInputNode* m_searchBar;
     CCLabelBMFont* m_countLabel;
     CCLabelBMFont* m_pageLabel;
+    CCMenuItemSpriteExtra* m_backButton;
+    CCMenuItemSpriteExtra* m_leftButton;
+    CCMenuItemSpriteExtra* m_rightButton;
+    CCMenuItemSpriteExtra* m_infoButton;
+    CCMenuItemSpriteExtra* m_refreshButton;
+    CCMenuItemSpriteExtra* m_pageButton;
+    CCMenuItemSpriteExtra* m_randomButton;
     CCMenuItemSpriteExtra* m_firstButton;
     CCMenuItemSpriteExtra* m_lastButton;
     int m_page = 0;
@@ -52,7 +48,8 @@ protected:
     bool init() override;
     void addSearchBar();
     void populateList(std::string query);
-    void loadLevelsFinished(CCArray* levels, const char*) override;
+    int getMaxPage();
+    void loadLevelsFinished(CCArray*, const char*) override;
     void loadLevelsFailed(const char*) override;
     void loadLevelsFinished(CCArray* levels, const char* key, int) override {
         loadLevelsFinished(levels, key);