+++ src/main.cpp
@@ -4,10 +4,11 @@
 class $modify(IDMenuLayer, MenuLayer) {
     bool init() {
         if (!MenuLayer::init()) return false;
-        if (!IDListLayer::AREDL_TRIED_LOADING) {
-            IDListLayer::AREDL_TRIED_LOADING = true;
-            IDListLayer::loadAREDL(true);
-        }
+
+        if (IDListLayer::AREDL_TRIED_LOADING) return true;
+        IDListLayer::AREDL_TRIED_LOADING = true;
+        IDListLayer::loadAREDL(true);
+
         return true;
     }
 };
@@ -19,19 +20,15 @@ class $modify(IDLevelSearchLayer, LevelSearchLayer) {
 
         auto demonlistButtonSprite = CircleButtonSprite::createWithSpriteFrameName("diffIcon_10_btn_001.png", 1.0f, CircleBaseColor::Pink, CircleBaseSize::Medium);
         demonlistButtonSprite->setScale(0.8f);
-        auto demonlistButton = CCMenuItemSpriteExtra::create(demonlistButtonSprite, this, menu_selector(IDLevelSearchLayer::onDemonList));
+        auto demonlistButton = CCMenuItemExt::createSpriteExtra(demonlistButtonSprite, [](auto) {
+            CCDirector::sharedDirector()->replaceScene(CCTransitionFade::create(0.5f, IDListLayer::scene()));
+        });
         demonlistButton->setID("demonlist-button"_spr);
         auto menu = getChildByID("other-filter-menu");
         menu->addChild(demonlistButton);
         menu->updateLayout();
 
         return true;
-    }
-
-    void onDemonList(CCObject*) {
-        auto scene = CCScene::create();
-        scene->addChild(IDListLayer::create());
-        CCDirector::sharedDirector()->replaceScene(CCTransitionFade::create(0.5f, scene));
     }
 };
 
@@ -48,19 +45,19 @@ class $modify(IDLevelCell, LevelCell) {
             auto found = std::find(begin, end, m_level->m_levelID.value());
             if (found != end) {
                 auto rankTextNode = CCLabelBMFont::create(fmt::format("#{} on AREDL", IDListLayer::AREDL_POSITIONS[found - begin]).c_str(), "chatFont.fnt");
-                rankTextNode->setPosition(346.0f, m_compactView ? 9.0f : 12.0f);
+                rankTextNode->setPosition(346.0f, m_level->m_dailyID.value() > 0 ? 6.0f : m_compactView ? 9.0f : 12.0f);
                 rankTextNode->setAnchorPoint({ 1.0f, 1.0f });
                 rankTextNode->setScale(m_compactView ? 0.45f : 0.6f);
-                rankTextNode->setColor({ 51, 51, 51 });
-                rankTextNode->setOpacity(152);
-                rankTextNode->setID("level-rank-label"_spr);
-                m_mainLayer->addChild(rankTextNode);
-                if (m_level->m_dailyID > 0 || Mod::get()->getSettingValue<bool>("white-rank")) {
+                if (m_level->m_dailyID.value() > 0 || Mod::get()->getSettingValue<bool>("white-rank")) {
                     rankTextNode->setColor({ 255, 255, 255 });
                     rankTextNode->setOpacity(200);
                 }
-
-                if (m_level->m_dailyID > 0) rankTextNode->setPositionY(6);
+                else {
+                    rankTextNode->setColor({ 51, 51, 51 });
+                    rankTextNode->setOpacity(152);
+                }
+                rankTextNode->setID("level-rank-label"_spr);
+                m_mainLayer->addChild(rankTextNode);
             }
         }
     }
@@ -71,7 +68,7 @@ class $modify(IDKeyboardDispatcher, CCKeyboardDispatcher) {
     bool dispatchKeyboardMSG(enumKeyCodes key, bool down, bool repeat) {
         auto layer = CCDirector::sharedDirector()->getRunningScene()->getChildByID("IDListLayer"_spr);
         if (key == KEY_Enter && down && layer) {
-            static_cast<IDListLayer*>(layer)->onSearch(nullptr);
+            static_cast<IDListLayer*>(layer)->search();
             return true;
         }
         else return CCKeyboardDispatcher::dispatchKeyboardMSG(key, down, repeat);